Reorganização dos módulos do projeto
