+++ persistence/tests/conftest.py
@@ -1,11 +1,11 @@
 from pyramid import testing
 import pytest
 
-from catalog_persistence.databases import (
+from persistence.databases import (
     CouchDBManager,
     InMemoryDBManager,
 )
-from catalog_persistence.services import DatabaseService
+from persistence.services import DatabaseService
 
 
 @pytest.yield_fixture
